# web/docs/openapi.yaml
openapi: 3.0.3
info:
  title: MagicStrike API
  description: |
    API para análise tática inteligente de replays de CS2 (Counter-Strike 2).
    
    ### Fluxo de Autenticação (Passwordless)
    1. O usuário solicita um Magic Link através de `/api/v1/auth/magic-link`.
    2. Um e-mail contendo um token seguro de utilização única é enviado (em ambiente de desenvolvimento/teste, o token é registrado nos logs/banco).
    3. O usuário realiza a requisição POST `/api/v1/auth/callback` com o token recebido.
    4. O servidor responde com um JWT (`access_token`) válido por 7 dias e um `session_id`.
    5. O frontend deve incluir o cabeçalho `Authorization: Bearer <access_token>` em todas as rotas protegidas.
    6. Se o usuário deslogar via `DELETE /api/v1/auth/session`, o token de sessão é revogado imediatamente e inserido em uma blocklist em memória com TTL correspondente à expiração do JWT.
    
    ### Upload de Demos (Replays)
    1. O frontend envia metadados da partida e o nome do arquivo para `/api/v1/demos/upload-request`.
    2. O backend responde com uma URL pré-assinada temporária para o MinIO/S3 (`upload_url`) e a chave do bucket correspondente (`bucket_key`).
    3. O frontend faz uma chamada HTTP PUT direta para a `upload_url` enviando o arquivo `.dem` binário (`Content-Type: application/octet-stream`).
...
  version: 1.0.0
servers:
  - url: http://localhost:8080/api/v1
    description: Servidor de Desenvolvimento Local
paths:
  /auth/magic-link:
    post:
      summary: Solicitar Magic Link para login
      description: Solicita o envio de um link mágico de autenticação para o e-mail informado. Para evitar enumeração de contas, o endpoint retorna sempre status 202 (Accepted).
      operationId: requestMagicLink
      tags:
        - Autenticação
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
              properties:
                email:
                  type: string
                  format: email
                  maxLength: 254
                  example: user@example.com
      responses:
        '202':
          description: Solicitação aceita. Se o e-mail existir, o link mágico será enviado.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: If the email exists, a magic link has been sent.
        '400':
          description: Corpo da requisição inválido ou e-mail ausente.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseBasic'
...
  /auth/refresh:
    post:
      summary: Renovar JWT e Sessão de Autenticação
      description: Gera um novo JWT com expiração renovada a partir de uma sessão ativa válida. Requer o cabeçalho Authorization.
      operationId: refreshSession
      tags:
        - Autenticação
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Sessão renovada com sucesso.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthSuccessResponse'
        '401':
          description: Não autorizado ou sessão expirada/inválida.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseBasic'
...
  /matches:
    get:
      summary: Listar partidas do usuário
      description: Retorna uma lista paginada das partidas enviadas pelo usuário logado.
      operationId: listMatches
      tags:
        - Partidas (Matches)
      security:
        - BearerAuth: []
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 50
            default: 20
          description: Quantidade máxima de registros retornados.
        - name: offset
          in: query
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
          description: Ponto de partida para a paginação.
      responses:
        '200':
          description: Lista de partidas retornada com sucesso.
          content:
            application/json:
              schema:
                type: object
                required:
                  - success
                  - data
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    required:
                      - matches
                      - limit
                      - offset
                      - count
                    properties:
                      matches:
                        type: array
                        items:
                          $ref: '#/components/schemas/Match'
                      limit:
                        type: integer
                        example: 20
                      offset:
                        type: integer
                        example: 0
                      count:
                        type: integer
                        example: 1
        '400':
          description: Parâmetros limit/offset inválidos.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseBasic'
        '401':
          description: Não autenticado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseBasic'
  /matches/{id}:
    get:
      summary: Obter detalhes de uma partida
      description: Retorna os dados táticos processados e metadados de uma partida específica de propriedade do usuário.
      operationId: getMatch
      tags:
        - Partidas (Matches)
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: ID único da partida.
      responses:
        '200':
          description: Partida encontrada.
          content:
            application/json:
              schema:
                type: object
                required:
                  - success
                  - data
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/Match'
        '401':
          description: Autenticação pendente.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseBasic'
        '404':
          description: Partida não encontrada ou não pertencente ao usuário.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseBasic'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    User:
      type: object
      required:
        - id
        - username
        - email
        - points
        - blocked
        - created_at
        - updated_at
      properties:
        id:
          type: string
          description: ID do usuário (gerado como ULID).
          example: 01KTAM9NZ3QF5V2N9H62R05JM7
        username:
          type: string
          example: "insanifan"
        email:
          type: string
          format: email
          example: "test@example.com"
        avatar:
          type: string
          format: uri
          example: "https://example.com/avatar.jpg"
        points:
          type: integer
          example: 0
        blocked:
          type: boolean
          example: false
        created_at:
          type: string
          format: date-time
          example: "2026-06-04T22:31:30-03:00"
        updated_at:
          type: string
          format: date-time
          example: "2026-06-04T22:31:30-03:00"
    AuthSuccessResponse:
      type: object
      required:
        - success
        - data
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          required:
            - access_token
            - session_id
            - user
            - expires_at
          properties:
            access_token:
              type: string
              description: Token JWT para autenticação.
              example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
            session_id:
              type: string
              description: ID da sessão criada.
              example: 01KTAP0QYEDTTHKHDCA1BMD26W
            user:
              $ref: '#/components/schemas/User'
            expires_at:
              type: string
              format: date-time
              example: "2026-06-12T01:23:24Z"
    Match:
      type: object
      required:
        - id
        - user_id
        - status
        - created_at
        - updated_at
      properties:
        id:
          type: string
          description: ID da partida (ULID).
          example: 01KTAP1DS10RT0VRSBD62Q0MQB
        user_id:
          type: string
          description: ID do proprietário da partida.
          example: 01KTAM9NZ3QF5V2N9H62R05JM7
        status:
          type: string
          enum: [pending, processing, processed, failed]
          description: Estado atual do parsing e processamento da demo.
          example: processed
        team_a:
          type: string
          example: MIBR
        team_b:
          type: string
          example: Lynn Vision
        demo_md5:
          type: string
          example: 5d41402abc4b2a76b9719d911017c592
        score_a:
          type: integer
          example: 13
        score_b:
          type: integer
          example: 6
        total_rounds:
          type: integer
          example: 19
        map_name:
          type: string
          example: de_anubis
        created_at:
          type: string
          format: date-time
          example: "2026-06-05T01:23:24Z"
        updated_at:
          type: string
          format: date-time
          example: "2026-06-05T01:25:00Z"
...
    ErrorResponseBasic:
      type: object
      required:
        - success
        - message
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: "Email is required"
...